Move StewardPlugin in new plugin subproject
With this change StewardPlugin.scala is compiled as part of the sbt
build and IDEs recognize this file now as proper source file.
